+++ src/constants.js
@@ -147,13 +147,13 @@ export const actionMenuDisplayPerms = [
   'ui-inventory.item.edit',
   'ui-inventory.item.delete',
   'ui-requests.create',
-  'ui-inventory.items.mark-items-withdrawn',
-  'ui-inventory.items.mark-intellectual-item',
-  'ui-inventory.items.mark-restricted',
-  'ui-inventory.items.mark-unknown',
-  'ui-inventory.items.mark-unavailable',
-  'ui-inventory.items.mark-long-missing',
-  'ui-inventory.items.mark-in-process-non-requestable',
+  'ui-inventory.items.mark-withdrawn.execute',
+  'ui-inventory.items.mark-intellectual-item.execute',
+  'ui-inventory.items.mark-restricted.execute',
+  'ui-inventory.items.mark-unknown.execute',
+  'ui-inventory.items.mark-unavailable.execute',
+  'ui-inventory.items.mark-long-missing.execute',
+  'ui-inventory.items.mark-in-process-non-requestable.execute',
 ];
 
 export const layers = {